I removed four chiral emoticons, ):  );  ]: and ]; because they almost always occur as the end of a parenthetical statement, e.g.
    You know her (or you want to): Miranda July

I now remove some of the html tags inside of posts, which are marked by &lt; &gt; stuff, e.g.

    &lt;a href="http://mirandajuly.com/ "&gt;Miranda July&lt;/a&gt; -> Miranda July
    &lt;b&gt;Bearcub48 the troll&lt;/b&gt; -> Bearcub48 the troll

But not if they enclose an ampersand or '<' character, e.g.

    &lt;</string>...stuff<string> Donnie is way &gt; me at just about everything
    &lt;&amp;lt; stuff &amp;gt;&gt;

To reduce the noise in the emoticons, increase recognition of common xml entities which end in a semicolon:
-Discarding line breaks: “&#xD;”
-Discarding “nonbreaking space” entities: "&nbsp;" or "&amp;nbsp;"
-Discarding quote entities: "&quot;" or "&amp;quot;"

Wherever and http:// or https:// shows up, this was generating a :/ emoticon. Now fixed these to be their own tokens.

Heart tokens "<3", "&lt;3", or "&amp;lt;3" are now tokenized (as the same token). Also broken hearts like &lt;/3 or <\3

If not used in a heart or html tag, "&lt;" "&gt;" or their alternatives "&amp;gt;" "&amp;lt;" are tokenized

All remaining ampersands "&amp;" or "&amp;amp;" are discarded after the above are processed

When matching ;</string>, return the punctuation to the stream so that e.g. (-:</string> is properly parsed.